+++ SeguroVeiculos/SeguroVeiculos.API/Models/AddSeguro/AddSeguroInput.cs
@@ -4,7 +4,7 @@
     {
         public string Nome { get; set; } = string.Empty;
         public string CPF { get; set; } = string.Empty;
-        public decimal ValorVeiculo { get; set; } = 0;
+        public double ValorVeiculo { get; set; } = 0;
         public string MarcaModeloVeiculo { get; set; } = string.Empty;
 
     }
